swapped torch.float to torch.bfloat

+++ generation/en/en_mistral7b.py
@@ -27,7 +27,7 @@ def answer_trivia(
     model = AutoModelForCausalLM.from_pretrained(
         model_name,
         gguf_file=file_name,
-        torch_dtype=torch.float8,
+        torch_dtype=torch.bfloat16,
         #attn_implementation="eager",
         #device_map="auto",
         #token = access_token
